backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "python-multipart<1.0.0,>=0.0.7",
    "email-validator<3.0.0.0,>=2.1.0.post1",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "tenacity<9.0.0,>=8.2.3",
    "pydantic>2.0",
    "emails<1.0,>=0.6",
    "jinja2<4.0.0,>=3.1.4",
    "alembic<2.0.0,>=1.12.1",
    "httpx<1.0.0,>=0.25.1",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "sqlmodel<1.0.0,>=0.0.21",
    # Pin bcrypt until passlib supports the latest
    "bcrypt==4.3.0",
    "pydantic-settings<3.0.0,>=2.2.1",
    "sentry-sdk[fastapi]<2.0.0,>=1.40.6",
    "pyjwt<3.0.0,>=2.8.0",
    "requests>=2.31.0,<3.0.0",
    "trafilatura>=1.6.0,<2.0.0",
    "extruct>=0.16.0,<1.0.0",
    "w3lib>=2.1.0,<3.0.0",
    "lxml>=4.9.0,<6.0.0",
    "html5lib>=1.1,<2.0",
    "beautifulsoup4>=4.12.0,<5.0.0",
    "jmespath>=1.0.1,<2.0.0",
    "PyYAML>=6.0.0,<7.0.0",
    # keybert and sentence-transformers are provided via the optional [project.optional-dependencies].ai group
    "redis>=5.0.0,<6.0.0",
    "rq>=1.15.0,<2.0.0",
    "prometheus-client>=0.19.0,<1.0.0",
    "prometheus-fastapi-instrumentator>=6.1.0,<7.0.0",
    "stripe>=5.0.0,<6.0.0",
    # Heavy AI deps moved to optional extras for faster local setup
    # Install via: uv pip install .[ai]  or  pip install ".[ai]"
]

[project.optional-dependencies]
ai = [
    "keybert>=0.8.0,<1.0.0",
    "sentence-transformers>=2.2.0,<3.0.0",
    "crewai>=0.1.0",
    "litellm>=0.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.coverage.run]
source = ["app"]
dynamic_context = "test_function"

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true

